+++ stories/flow-core/f-checkbox.stories.js
@@ -1,6 +1,6 @@
+import { useArgs, useState } from "storybook/preview-api";
 import { html } from "lit-html";
 import { unsafeSVG } from "lit-html/directives/unsafe-svg.js";
-import { useArgs, useEffect, useState } from "@storybook/client-api";
 import fCheckboxAnatomy from "../svg/i-fcheckbox-anatomy.js";
 
 export default {