+++ stories/flow-lineage/node-overlap.stories.ts
@@ -1,4 +1,4 @@
-import { Story, Meta } from "@storybook/web-components";
+import { Story, Meta } from "@storybook/web-components-vite";
 import { html } from "lit-html";
 import { LineageNodeLinks } from "@nonfx/flow-lineage";