+++ packages/flow-core/CHANGELOG.md
@@ -1,5 +1,13 @@
 <h4 className="margin-btm-8">Release Notes</h4>
 
+## 5.1.1
+
+### Patch Changes
+
+- Fixes
+- Updated dependencies
+  - @nonfx/flow-core-config@1.3.2
+
 ## 5.1.0
 
 ### Minor Changes